[NEW] Make libraries and headers installable.
To create an executable linking to libyijinjing.so we need to include
all headers, .a and .so libraries to proper folder structure. This
commit adds proper path configurations in CMake.

The commit also includes an examples/ folder which builds a
yijinjing-demo application and link to libyijinjing.so.

Note that for a purpose of keeping devbox environment clean, our headers
and libraries are saved to ~/.local/kungfu (build-cxx-core.sh).
Thus, we need to define LD_LIBRARY_PATH=$HOME/.local/kungfu/lib
environment variable to run yijinjing-demo executable. See
build-run-cxx-examples.sh for details.